+++ lib/easy_ads_flutter.dart
@@ -5,6 +5,7 @@ export 'package:easy_ads_flutter/src/enums/ad_unit_type.dart';
 export 'package:easy_ads_flutter/src/enums/ad_event_type.dart';
 export 'package:easy_ads_flutter/src/utils/ad_event.dart';
 export 'package:easy_ads_flutter/src/utils/i_ad_id_manager.dart';
+export 'package:easy_ads_flutter/src/utils/easy_banner_ad.dart';
 export 'package:easy_ads_flutter/src/easy_admob/easy_admob_banner_ad.dart';
 export 'package:easy_ads_flutter/src/easy_unity/easy_unity_banner_ad.dart';
 export 'package:easy_ads_flutter/src/easy_ads.dart';