+++ lib/src/utils/i_ad_id_manager.dart
@@ -1,17 +1,33 @@
 abstract class IAdIdManager {
   const IAdIdManager();
 
+  /// Pass null if you wish not to implement unity ads
+  ///
+  /// AppAdIds? get unityAdIds => null;
   AppAdIds? get unityAdIds;
+
+  /// Pass null if you wish not to implement admob ads
+  ///
+  /// AppAdIds? get admobAdIds => null;
   AppAdIds? get admobAdIds;
-  AppAdIds? get fbAdIds;
+
+  /// Pass null if you wish not to implement appLovin ads
+  ///
+  /// AppAdIds? get appLovinAdIds => null;
   AppAdIds? get appLovinAdIds;
 }
 
 class AppAdIds {
+  /// App Id should never be null, if there is no app id for a particular ad network, leave it empty
   final String appId;
 
+  /// if id is null, it will not be implemented
   final String? interstitialId;
+
+  /// if id is null, it will not be implemented
   final String? rewardedId;
+
+  /// if id is null, it will not be implemented
   final String? bannerId;
 
   const AppAdIds({